+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog of lizard-fewsnorm
 0.7.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Load and dump using natural keys now works, except for
+  GeoLocationCache because it is inherited from another model.
 
 
 0.7.1 (2012-01-31)